v20

+++ main.py
@@ -5,15 +5,14 @@ from models.account import Account
 from models.transaction import Transaction
 from services.transaction_management import create_id_for_transaction, generate_report_user_story_6
 from services.transaction_management import generate_report_user_story_7
-from services.account_management import create_account_object_from_json, update_account
-from services.account_management import create_account, isValid, account_proof
+from services.account_management import create_account_object_from_json, update_account, is_correct_amount
+from services.account_management import create_account, is_correct_email, account_proof
 from services.authentication import login_fnc, logout_fnc, user_mail_from_Json
 from services.authentication import update_user_information
 from utils.currency_converter import converter
 from config.settings import Constants_and_variables
 
 print("\n")
-# print("*****************************************")
 print("₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽₽")
 print("€€€            ~~~~~~~~~~               €€€")
 print("$$$    FINANCIAL ACCOUNTING PROGRAM©    $$$")
@@ -32,6 +31,8 @@ while not username:
     else:
         print("\nLogin or password is wrong!\nWant to try again?\n")
         response = input("If yes, print Y,\nif You want to register, print R,\nif Exit, print any another key: ")
+
+        # Registration new user
         if response == "R" or response == "r":
             print("Please register\n")
             name = input("Enter your name: ")
@@ -39,7 +40,7 @@ while not username:
             login = name + surname
             while True:
                 email = input("Enter your email: ")
-                if not isValid(email):
+                if not is_correct_email(email):
                     print("\nEmail is invalid!\n")
                 else:
                     break
@@ -61,12 +62,9 @@ while not username:
                     print(err)
 
             while True:
-                try:
-                    balance = round(float(input("Enter the start balance in your account: ")), 2)
+                balance = input("Enter the amount: ")
+                if is_correct_amount(balance):
                     break
-                except ValueError:
-                    print("Invalid balance. Please try again.")
-
             create_account(username, name, currency, balance)
 
             print("")
@@ -96,13 +94,12 @@ while True:
         except ValueError:
             print("Invalid choice. Please try again.")
 
-
     # Update_profile
     if choice == 1:
         password = input("Enter new password: ")
         while True:
             email = input("Enter new email: ")
-            if not isValid(email):
+            if not is_correct_email(email):
                 print("\nInvalid email forma!\n")
             else:
                 break
@@ -139,43 +136,58 @@ while True:
         transaction_id = create_id_for_transaction(username)
 
         # Запрашиваем у пользователя данные о транзакции
-        account_id = account_proof(username, "Enter your account ID: ")
-        amount = float(input("Enter the amount: "))
+        while True:
+            account_id = account_proof(username, "Enter your account ID: ")
+            if not account_id:
+                print("Try again!")
+            else:
+                break
+        while True:
+            amount = input("Enter the amount: ")
+            if is_correct_amount(amount):
+                break
         transaction_type = input("Enter the type of transaction\n1 - Adding income \n2 - Expense registration: ")
-        if transaction_type == "1":
-            transaction_type = "Income"
-        elif transaction_type == "2":
-            transaction_type = "Payment"
-        else:
-            print("Not correct number, please try again")
-            break
+        while True:
+            if transaction_type == "1":
+                transaction_type = "Income"
+                break
+            elif transaction_type == "2":
+                transaction_type = "Payment"
+                break
+            else:
+                print("Not correct number, please try again")
+
         description = input("Write a description of transaction: ")
         print("\n")
+
+        print(transaction_id, username, account_id, amount, transaction_type, description,
+              datetime.datetime.now().strftime('%d.%m.%Y %H:%M:%S'))
 
         # Create object of klass Transaction
         transaction = Transaction(transaction_id, username, account_id, amount, transaction_type, description,
                                   datetime.datetime.now().strftime('%d.%m.%Y %H:%M:%S'))
-
         transaction.record_transaction()
 
     # Display balance for account
     elif choice == 4:
         account_id = account_proof(username, "Please, enter your account ID: ")
-        account = create_account_object_from_json(username, account_id)
-        result = account.get_balance()
-        print(f'Balance {account.name}: {account.get_balance()} {Constants_and_variables.currency[account.currency]}')
+        if account_id:
+            account = create_account_object_from_json(username, account_id)
+            result = account.get_balance()
+            print(f'Balance {account.name}: {account.get_balance()} {Constants_and_variables.currency[account.currency]}')
 
 
     # Delete account
     elif choice == 5:
         account_id = account_proof(username, "Please enter the ID\nwhich you want to delete: ")
-        user_response = input("Are You sure?: Υ or N: ")
-        if user_response in ("Y", "y"):
-            account = create_account_object_from_json(username, account_id)
-            account.delete_account()
-            print(f"Account: {account.name} successfully deleted!\n")
-        elif user_response in ("N", "n"):
-            print("Continue work!\n")
+        if account_id:
+            user_response = input("Are You sure?: Υ or N: ")
+            if user_response in ("Y", "y"):
+                account = create_account_object_from_json(username, account_id)
+                account.delete_account()
+                print(f"Account: {account.name} successfully deleted!\n")
+            elif user_response in ("N", "n"):
+                print("Continue work!\n")
 
 
     # Report user story 6
@@ -235,12 +247,16 @@ while True:
                     currency = int(input(
                         "Enter the number of currency for your account\n 1 - rub \n 2 - $ \n 3 - €\n 4 - kzt\n 5 - cny\n 6 - byn\n Your choice: "))
                     if currency not in (1, 2, 3, 4, 5, 6):
-                        print("Invalid choice. Please try again.")
+                        print("Invalid choice. Please try again\n")
                     else:
                         break
                 except ValueError as err:
                     print(err)
-            balance = round(float(input("Enter the new account balance: ")), 2)
+            while True:
+                amount = input("Enter the amount: ")
+                if is_correct_amount(amount):
+                    break
+
             update_account(username, account_id, name, currency, balance)
             print(f"Account: {name} successfully updated!\n")
         elif user_response in ("N", "n"):
@@ -250,17 +266,28 @@ while True:
 
     # Transfer
     elif choice == 9:
-        account_id = account_proof(username, "Enter the account ID from which you want to transfer: ")
-        account_receiver = account_proof(username, "Enter the account ID to which you want to transfer: ")
+        # account_id = account_proof(username, "Enter the account ID from which you want to transfer: ")
+        while True:
+            account_id = account_proof(username, "Enter the account ID from which you want to transfer: ")
+            if not account_id:
+                print("Try again!")
+            else:
+                break
+        while True:
+            account_receiver = account_proof(username, "Enter the account ID from which you want to transfer: ")
+            if not account_id:
+                print("Try again!")
+            else:
+                break
+
+        # account_receiver = account_proof(username, "Enter the account ID to which you want to transfer: ")
         if account_id == account_receiver:
             print("Accounts ID for transfer must be different!")
             break
         while True:
-            try:
-                amount = round(float(input("Enter the amount: ")), 2)
+            amount = input("Enter the amount: ")
+            if is_correct_amount(amount):
                 break
-            except ValueError:
-                print("Not correct number, please try again")
         account = create_account_object_from_json(username, str(account_id))
         account.transfer(str(account_receiver), amount)
 
@@ -290,11 +317,9 @@ while True:
                 print(err)
 
         while True:
-            try:
-                amount = round(float(input("Enter the amount: ")), 2)
+            amount = input("Enter the amount: ")
+            if is_correct_amount(amount):
                 break
-            except ValueError:
-                print("Invalid balance. Please try again.")
         res = asyncio.run(
             converter(Constants_and_variables.currency[currency_first], Constants_and_variables.currency[currency_two],
                       amount))
